Reviewer #1
When I run the code initially, it pops up a type error window, "setting a property that has only a getter at /index.html (line 809)". I tried, unsuccessfully, to locate line 809. If I dismiss this error, the code runs fine, so it's not a high priority issue. 

There are two windows with a navigation between the two. The code seems to run as intended. 

I'm unable to run the app on multiple devices, but when I change the size of my web screen, the text seems to fill the space proportionally, so it adjusts to different sized screens.